Update README.md
went off of the following:

The ReadMe.md should include the following:

Brief Application description
Contributors - In this case, it will be the team Members
Technology Stack used for the project
Prerequisites to run the application - Any software that needs to be installed to run the application
Instructions on how to run the application locally.
How to run the tests
Link to the deployed application
You should be updating the ReadMe.md file as you progress through the project. Remember to add/revise sections as you add new features to the application.